// src/app/home/models/track.model.ts
import { ExternalUrls } from './externalurls.model';
import { Artist } from './artist.model';
import { Album } from './album.model';
import { ExternalIds } from './externalids.model';

export interface Track {
    album: Album;
    artists: Artist[];
    available_markets: string[];
    disc_number: number;
    duration_ms: number;
    explicit: boolean;
    external_ids: ExternalIds;
    external_urls: ExternalUrls;
    href: string;
    id: string;
    is_local: boolean;
    name: string;
    popularity: number;
    preview_url: string;
    track_number: number;
    type: string;
    uri: string;
}